Revert "FFM-11385 Re-generate client API to bring in target-segment rules"
This reverts commit 9fca2c895318a4d953901b0141ac670872d76bad.
