+++ featureflags/openapi/config/api/client/get_evaluations.py
@@ -5,6 +5,7 @@ import httpx
 
 from ... import errors
 from ...client import AuthenticatedClient, Client
+from ...models.get_evaluations_response_200 import GetEvaluationsResponse200
 from ...types import UNSET, Response, Unset
 
 
@@ -34,7 +35,11 @@ def _get_kwargs(
 
 def _parse_response(
     *, client: Union[AuthenticatedClient, Client], response: httpx.Response
-) -> Optional[Any]:
+) -> Optional[GetEvaluationsResponse200]:
+    if response.status_code == HTTPStatus.OK:
+        response_200 = GetEvaluationsResponse200.from_dict(response.json())
+
+        return response_200
     if client.raise_on_unexpected_status:
         raise errors.UnexpectedStatus(response.status_code, response.content)
     else:
@@ -43,7 +48,7 @@ def _parse_response(
 
 def _build_response(
     *, client: Union[AuthenticatedClient, Client], response: httpx.Response
-) -> Response[Any]:
+) -> Response[GetEvaluationsResponse200]:
     return Response(
         status_code=HTTPStatus(response.status_code),
         content=response.content,
@@ -58,7 +63,7 @@ def sync_detailed(
     *,
     client: AuthenticatedClient,
     cluster: Union[Unset, str] = UNSET,
-) -> Response[Any]:
+) -> Response[GetEvaluationsResponse200]:
     """Get feature evaluations for target
 
     Args:
@@ -71,7 +76,7 @@ def sync_detailed(
         httpx.TimeoutException: If the request takes longer than Client.timeout.
 
     Returns:
-        Response[Any]
+        Response[GetEvaluationsResponse200]
     """
 
     kwargs = _get_kwargs(
@@ -87,13 +92,13 @@ def sync_detailed(
     return _build_response(client=client, response=response)
 
 
-async def asyncio_detailed(
+def sync(
     environment_uuid: str,
     target: str,
     *,
     client: AuthenticatedClient,
     cluster: Union[Unset, str] = UNSET,
-) -> Response[Any]:
+) -> Optional[GetEvaluationsResponse200]:
     """Get feature evaluations for target
 
     Args:
@@ -106,7 +111,37 @@ async def asyncio_detailed(
         httpx.TimeoutException: If the request takes longer than Client.timeout.
 
     Returns:
-        Response[Any]
+        GetEvaluationsResponse200
+    """
+
+    return sync_detailed(
+        environment_uuid=environment_uuid,
+        target=target,
+        client=client,
+        cluster=cluster,
+    ).parsed
+
+
+async def asyncio_detailed(
+    environment_uuid: str,
+    target: str,
+    *,
+    client: AuthenticatedClient,
+    cluster: Union[Unset, str] = UNSET,
+) -> Response[GetEvaluationsResponse200]:
+    """Get feature evaluations for target
+
+    Args:
+        environment_uuid (str):
+        target (str):
+        cluster (Union[Unset, str]):
+
+    Raises:
+        errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
+        httpx.TimeoutException: If the request takes longer than Client.timeout.
+
+    Returns:
+        Response[GetEvaluationsResponse200]
     """
 
     kwargs = _get_kwargs(
@@ -118,3 +153,35 @@ async def asyncio_detailed(
     response = await client.get_async_httpx_client().request(**kwargs)
 
     return _build_response(client=client, response=response)
+
+
+async def asyncio(
+    environment_uuid: str,
+    target: str,
+    *,
+    client: AuthenticatedClient,
+    cluster: Union[Unset, str] = UNSET,
+) -> Optional[GetEvaluationsResponse200]:
+    """Get feature evaluations for target
+
+    Args:
+        environment_uuid (str):
+        target (str):
+        cluster (Union[Unset, str]):
+
+    Raises:
+        errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
+        httpx.TimeoutException: If the request takes longer than Client.timeout.
+
+    Returns:
+        GetEvaluationsResponse200
+    """
+
+    return (
+        await asyncio_detailed(
+            environment_uuid=environment_uuid,
+            target=target,
+            client=client,
+            cluster=cluster,
+        )
+    ).parsed
