--- featureflags/openapi/config/api/metrics/post_metrics.py
@@ -1,140 +1,0 @@
-from http import HTTPStatus
-from typing import Any, Dict, Optional, Union
-
-import httpx
-
-from ... import errors
-from ...client import AuthenticatedClient, Client
-from ...models.metrics import Metrics
-from ...types import UNSET, Response, Unset
-
-
-def _get_kwargs(
-    environment_uuid: str,
-    *,
-    body: Metrics,
-    cluster: Union[Unset, str] = UNSET,
-) -> Dict[str, Any]:
-    headers: Dict[str, Any] = {}
-
-    params: Dict[str, Any] = {}
-
-    params["cluster"] = cluster
-
-    params = {k: v for k, v in params.items() if v is not UNSET and v is not None}
-
-    _kwargs: Dict[str, Any] = {
-        "method": "post",
-        "url": "/metrics/{environment_uuid}".format(
-            environment_uuid=environment_uuid,
-        ),
-        "params": params,
-    }
-
-    _body = body.to_dict()
-
-    _kwargs["json"] = _body
-    headers["Content-Type"] = "application/json"
-
-    _kwargs["headers"] = headers
-    return _kwargs
-
-
-def _parse_response(
-    *, client: Union[AuthenticatedClient, Client], response: httpx.Response
-) -> Optional[Any]:
-    if response.status_code == HTTPStatus.OK:
-        return None
-    if response.status_code == HTTPStatus.UNAUTHORIZED:
-        return None
-    if response.status_code == HTTPStatus.FORBIDDEN:
-        return None
-    if response.status_code == HTTPStatus.INTERNAL_SERVER_ERROR:
-        return None
-    if client.raise_on_unexpected_status:
-        raise errors.UnexpectedStatus(response.status_code, response.content)
-    else:
-        return None
-
-
-def _build_response(
-    *, client: Union[AuthenticatedClient, Client], response: httpx.Response
-) -> Response[Any]:
-    return Response(
-        status_code=HTTPStatus(response.status_code),
-        content=response.content,
-        headers=response.headers,
-        parsed=_parse_response(client=client, response=response),
-    )
-
-
-def sync_detailed(
-    environment_uuid: str,
-    *,
-    client: AuthenticatedClient,
-    body: Metrics,
-    cluster: Union[Unset, str] = UNSET,
-) -> Response[Any]:
-    """Send metrics to the Analytics server.
-
-     Send metrics to Analytics server
-
-    Args:
-        environment_uuid (str):
-        cluster (Union[Unset, str]):
-        body (Metrics):
-
-    Raises:
-        errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
-        httpx.TimeoutException: If the request takes longer than Client.timeout.
-
-    Returns:
-        Response[Any]
-    """
-
-    kwargs = _get_kwargs(
-        environment_uuid=environment_uuid,
-        body=body,
-        cluster=cluster,
-    )
-
-    response = client.get_httpx_client().request(
-        **kwargs,
-    )
-
-    return _build_response(client=client, response=response)
-
-
-async def asyncio_detailed(
-    environment_uuid: str,
-    *,
-    client: AuthenticatedClient,
-    body: Metrics,
-    cluster: Union[Unset, str] = UNSET,
-) -> Response[Any]:
-    """Send metrics to the Analytics server.
-
-     Send metrics to Analytics server
-
-    Args:
-        environment_uuid (str):
-        cluster (Union[Unset, str]):
-        body (Metrics):
-
-    Raises:
-        errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
-        httpx.TimeoutException: If the request takes longer than Client.timeout.
-
-    Returns:
-        Response[Any]
-    """
-
-    kwargs = _get_kwargs(
-        environment_uuid=environment_uuid,
-        body=body,
-        cluster=cluster,
-    )
-
-    response = await client.get_async_httpx_client().request(**kwargs)
-
-    return _build_response(client=client, response=response)
