--- featureflags/openapi/config/api/proxy/authenticate_proxy_key.py
@@ -1,180 +1,0 @@
-from http import HTTPStatus
-from typing import Any, Dict, Optional, Union, cast
-
-import httpx
-
-from ... import errors
-from ...client import AuthenticatedClient, Client
-from ...models.authenticate_proxy_key_body import AuthenticateProxyKeyBody
-from ...models.authentication_response import AuthenticationResponse
-from ...types import Response
-
-
-def _get_kwargs(
-    *,
-    body: AuthenticateProxyKeyBody,
-) -> Dict[str, Any]:
-    headers: Dict[str, Any] = {}
-
-    _kwargs: Dict[str, Any] = {
-        "method": "post",
-        "url": "/proxy/auth",
-    }
-
-    _body = body.to_dict()
-
-    _kwargs["json"] = _body
-    headers["Content-Type"] = "application/json"
-
-    _kwargs["headers"] = headers
-    return _kwargs
-
-
-def _parse_response(
-    *, client: Union[AuthenticatedClient, Client], response: httpx.Response
-) -> Optional[Union[Any, AuthenticationResponse]]:
-    if response.status_code == HTTPStatus.OK:
-        response_200 = AuthenticationResponse.from_dict(response.json())
-
-        return response_200
-    if response.status_code == HTTPStatus.UNAUTHORIZED:
-        response_401 = cast(Any, None)
-        return response_401
-    if response.status_code == HTTPStatus.FORBIDDEN:
-        response_403 = cast(Any, None)
-        return response_403
-    if response.status_code == HTTPStatus.NOT_FOUND:
-        response_404 = cast(Any, None)
-        return response_404
-    if response.status_code == HTTPStatus.INTERNAL_SERVER_ERROR:
-        response_500 = cast(Any, None)
-        return response_500
-    if client.raise_on_unexpected_status:
-        raise errors.UnexpectedStatus(response.status_code, response.content)
-    else:
-        return None
-
-
-def _build_response(
-    *, client: Union[AuthenticatedClient, Client], response: httpx.Response
-) -> Response[Union[Any, AuthenticationResponse]]:
-    return Response(
-        status_code=HTTPStatus(response.status_code),
-        content=response.content,
-        headers=response.headers,
-        parsed=_parse_response(client=client, response=response),
-    )
-
-
-def sync_detailed(
-    *,
-    client: Union[AuthenticatedClient, Client],
-    body: AuthenticateProxyKeyBody,
-) -> Response[Union[Any, AuthenticationResponse]]:
-    """Endpoint that the Proxy can use to authenticate with the client server
-
-     Endpoint that the Proxy can use to authenticate with the client server
-
-    Args:
-        body (AuthenticateProxyKeyBody):
-
-    Raises:
-        errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
-        httpx.TimeoutException: If the request takes longer than Client.timeout.
-
-    Returns:
-        Response[Union[Any, AuthenticationResponse]]
-    """
-
-    kwargs = _get_kwargs(
-        body=body,
-    )
-
-    response = client.get_httpx_client().request(
-        **kwargs,
-    )
-
-    return _build_response(client=client, response=response)
-
-
-def sync(
-    *,
-    client: Union[AuthenticatedClient, Client],
-    body: AuthenticateProxyKeyBody,
-) -> Optional[Union[Any, AuthenticationResponse]]:
-    """Endpoint that the Proxy can use to authenticate with the client server
-
-     Endpoint that the Proxy can use to authenticate with the client server
-
-    Args:
-        body (AuthenticateProxyKeyBody):
-
-    Raises:
-        errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
-        httpx.TimeoutException: If the request takes longer than Client.timeout.
-
-    Returns:
-        Union[Any, AuthenticationResponse]
-    """
-
-    return sync_detailed(
-        client=client,
-        body=body,
-    ).parsed
-
-
-async def asyncio_detailed(
-    *,
-    client: Union[AuthenticatedClient, Client],
-    body: AuthenticateProxyKeyBody,
-) -> Response[Union[Any, AuthenticationResponse]]:
-    """Endpoint that the Proxy can use to authenticate with the client server
-
-     Endpoint that the Proxy can use to authenticate with the client server
-
-    Args:
-        body (AuthenticateProxyKeyBody):
-
-    Raises:
-        errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
-        httpx.TimeoutException: If the request takes longer than Client.timeout.
-
-    Returns:
-        Response[Union[Any, AuthenticationResponse]]
-    """
-
-    kwargs = _get_kwargs(
-        body=body,
-    )
-
-    response = await client.get_async_httpx_client().request(**kwargs)
-
-    return _build_response(client=client, response=response)
-
-
-async def asyncio(
-    *,
-    client: Union[AuthenticatedClient, Client],
-    body: AuthenticateProxyKeyBody,
-) -> Optional[Union[Any, AuthenticationResponse]]:
-    """Endpoint that the Proxy can use to authenticate with the client server
-
-     Endpoint that the Proxy can use to authenticate with the client server
-
-    Args:
-        body (AuthenticateProxyKeyBody):
-
-    Raises:
-        errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
-        httpx.TimeoutException: If the request takes longer than Client.timeout.
-
-    Returns:
-        Union[Any, AuthenticationResponse]
-    """
-
-    return (
-        await asyncio_detailed(
-            client=client,
-            body=body,
-        )
-    ).parsed
