--- featureflags/openapi/config/models/metrics_data_metrics_type.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-
-class MetricsDataMetricsType(str, Enum):
-    FFMETRICS = "FFMETRICS"
-
-    def __str__(self) -> str:
-        return str(self.value)
